tf-assets : la vraie géométrie du serveur, résolue sans un seul trou
Les benchs de maillage INVENTAIENT leurs cuboïdes. Ils avaient la bonne
distribution — mesurée sur le codex — mais pas la bonne forme, et aucun
chiffre de rendu ne vaut tant que la géométrie est fabriquée.

`tf-assets` lit un pack : `blockstates → models (parent) → textures`, chaîne
qu'on ne peut pas raccourcir puisque `grass_block` n'a aucune texture qui porte
son nom et qu'un bloc peut n'être pas un cube. Il ne redistribue RIEN — on lit
l'installation de l'utilisateur, comme WorldPainter et Litematica, et le
dossier d'un launcher contient aussi les packs du SERVEUR, donc les blocs
custom arrivent avec leur géométrie sans rien demander.

Sur le pack réel du serveur : **2 560 blockstates, 6 233 modèles résolus, ZÉRO
introuvable, en 254 ms.** Le recensement retrouve indépendamment les chiffres
qui dimensionnent la phase 2 — 68,7 % de blocs-modèles côté minefield — et le
pire cas attendu, `pumpkin_treat_bag` à 82 cuboïdes.

Bout à bout, sur un quart de région bâtie et la géométrie RÉELLE :

    états distincts   766, dont 0 hors du pack
    table de formes   1,6 ms
    mailler, 4 fils    53 ms
    quads gloutons    530 132
    poses de modèles  491 969
    pour le GPU       12,4 Mo

Cinq pièges, tous mesurés sur le pack plutôt que supposés :

- **Un cuboïde en seizièmes ENTIERS re-quantifie le pack.** 17,2 % des 125 382
  coordonnées de modèle ne sont pas entières, dont des DIXIÈMES qu'aucune
  fraction binaire ne représente. Les cuboïdes et les quads passent en
  flottants ; ceux de la passe gloutonne tombent sur des multiples de 16, donc
  restent exacts et comparables sans tolérance.
- **`uv` absent n'est pas `[0,0,0,0]`**, et 26 à 32 % des faces sont dans ce
  cas. Les uv se déduisent des bornes du cuboïde — c'est ce qui fait qu'une
  dalle montre la moitié basse de sa texture.
- **`cull` ne porte que les faces qui DÉCLARENT `cullface`.** Le dessus d'une
  dalle est à ras et sans cullface : un bloc posé dessus ne doit pas le faire
  disparaître.
- **« Un cuboïde » n'est pas le critère d'un cube.** `grass_block` en déclare
  deux ; le critère est un cuboïde qui REMPLIT, et à égalité celui qui déclare
  le plus de faces. Un test le vérifie sur un modèle à deux cuboïdes.
- **Un modèle nomme son parent, et un pack vient du disque d'un utilisateur.**
  `../../../etc/passwd` est un nom bien formé, et un cycle de parents boucle
  jusqu'à la mort du processus. Les deux sont refusés.

Ce que le pack ne dit PAS reste une couture explicite : le verre remplit son
bloc et ne doit masquer personne, et la réponse est dans l'alpha de sa texture.
`table_formes` prend donc un prédicat `translucide` plutôt qu'une liste de noms
en dur, qui ne couvrirait aucun bloc `minefield:*`.

Trois tests tournent contre le pack réel quand `TF_PACK` le désigne, et se
sautent sinon : un test qu'on ne peut pas jouer sans une donnée privée ne doit
pas faire échouer la suite de quelqu'un qui ne l'a pas, mais un pack écrit à la
main ne reproduit jamais ce qu'un vrai serveur contient.

298 tests.

+++ crates/tf-mesh/benches/maillage.rs
@@ -36,11 +36,11 @@ fn table() -> (TableFormes, Vec<StateId>) {
         let cub: Vec<Cuboide> = (0..*n)
             .map(|k| {
                 let t = (*n).max(1) as i32;
-                let bas = (k as i32 * 16 / t) as i8;
-                let haut = ((k as i32 + 1) * 16 / t).max(bas as i32 + 1).min(16) as i8;
+                let bas = (k as i32 * 16 / t) as f32;
+                let haut = (((k as i32 + 1) * 16 / t).max(bas as i32 + 1).min(16)) as f32;
                 Cuboide {
-                    min: [0, bas, 0],
-                    max: [16, haut, 16],
+                    min: [0.0, bas, 0.0],
+                    max: [16.0, haut, 16.0],
                     faces: 0x3F,
                     cull: 0x3F,
                 }
@@ -215,11 +215,12 @@ fn charger(octets: &[u8], b: &Build) -> (tf_mesh::Grille, TableFormes) {
                     (0..*n)
                         .map(|k| {
                             let tt = (*n).max(1) as i32;
-                            let bas = (k as i32 * 16 / tt) as i8;
-                            let haut = ((k as i32 + 1) * 16 / tt).max(bas as i32 + 1).min(16) as i8;
+                            let bas = (k as i32 * 16 / tt) as f32;
+                            let haut =
+                                (((k as i32 + 1) * 16 / tt).max(bas as i32 + 1).min(16)) as f32;
                             Cuboide {
-                                min: [0, bas, 0],
-                                max: [16, haut, 16],
+                                min: [0.0, bas, 0.0],
+                                max: [16.0, haut, 16.0],
                                 faces: 0x3F,
                                 cull: 0x3F,
                             }
