+++ crates/tf-mesh/src/maillage.rs
@@ -6,16 +6,18 @@ use crate::forme::Face;
 
 /// Un quad, en **seizièmes de bloc** dans le repère de la section.
 ///
-/// Des entiers et non des flottants : les modèles Minecraft sont déjà en
-/// seizièmes, la passe gloutonne travaille sur des bords de bloc, et un entier
-/// se compare exactement. Un quad de 3 × 5 blocs sort à 48 × 80.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// La même unité que les modèles, et en flottants pour la même raison : 17 %
+/// des coordonnées du pack Minefield ne sont pas entières. Les quads de la
+/// passe gloutonne, eux, tombent toujours sur des bords de bloc — donc sur des
+/// multiples de 16, exacts en flottant, et comparables sans tolérance. Un quad
+/// de 3 × 5 blocs sort à 48 × 80.
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub struct Quad {
     /// Coin de plus petites coordonnées, dans le repère de la section.
-    pub min: [i16; 3],
+    pub min: [f32; 3],
     /// Étendue dans les deux axes du plan de la face, dans l'ordre croissant
     /// des axes restants. Pour `±Y` : `[x, z]`.
-    pub taille: [i16; 2],
+    pub taille: [f32; 2],
     pub face: Face,
     /// L'état qui a produit ce quad — c'est lui qui désigne la tuile d'atlas.
     pub id: StateId,
@@ -24,8 +26,8 @@ pub struct Quad {
 impl Quad {
     /// Surface en seizièmes carrés. Sert aux tests de conservation : la somme
     /// des surfaces d'un maillage glouton doit égaler celle du maillage naïf.
-    pub fn aire(&self) -> i64 {
-        self.taille[0] as i64 * self.taille[1] as i64
+    pub fn aire(&self) -> f64 {
+        self.taille[0] as f64 * self.taille[1] as f64
     }
 }
 
@@ -55,7 +57,7 @@ impl Maillage {
         self.quads.is_empty()
     }
 
-    pub fn aire(&self) -> i64 {
+    pub fn aire(&self) -> f64 {
         self.quads.iter().map(Quad::aire).sum()
     }
 }
